Enable metric collection from arbitrary sources
