--- spec/vcr_cassettes/manageiq/providers/kubernetes/container_manager/metrics_capture/capture_context_container_timespan.yml
@@ -1,179 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://capture.context.com/hawkular/metrics/status
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      User-Agent:
-      - rest-client/2.0.2 (linux-gnu x86_64) ruby/2.4.1p111
-      Hawkular-Tenant:
-      - default
-      Authorization:
-      - Bearer theToken
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Encoding:
-      - gzip
-      Connection:
-      - keep-alive
-      X-Powered-By:
-      - Undertow/1
-      Server:
-      - WildFly/10
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '154'
-      Date:
-      - Thu, 22 Jun 2017 18:57:52 GMT
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        H4sIAAAAAAAAAA3MvQ7CIBiF4XthFsJfK3Fr1aqDizTulH5NSIA2QF2M9y7TWZ7zftETSnI2a0gfZwGdkB6713i9oAN6hM1DgFhMcWvEb0i5biWU8JYwMrhofHX97nzBQ1oDvrmC9b1jFclmYlJZJSTMLadMLawRlgumjpOUojUgOVUL1MDZ5GzinEy97Rv6/QHfcrqrlgAAAA==
-    http_version: 
-  recorded_at: Thu, 22 Jun 2017 18:57:52 GMT
-- request:
-    method: get
-    uri: https://capture.context.com/hawkular/metrics/counters/registry%2F476e8b1a-3a1a-11e7-a3eb-001a4a2314d7%2Fcpu%2Fusage/stats/?bucketDuration=20s&end=1498156842000&start=1498156522000
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      User-Agent:
-      - rest-client/2.0.2 (linux-gnu x86_64) ruby/2.4.1p111
-      Hawkular-Tenant:
-      - default
-      Authorization:
-      - Bearer theToken
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Encoding:
-      - gzip
-      Connection:
-      - keep-alive
-      X-Powered-By:
-      - Undertow/1
-      Server:
-      - WildFly/10
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '416'
-      Date:
-      - Thu, 22 Jun 2017 18:57:53 GMT
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        H4sIAAAAAAAAAI2UQWrDMBBF7+J1CNJIGknZ5xSlC0PcEohDiJ3SUnr3OrWE59sqaPuIyHvMxy/fzTC297E5aBuDduyIlFK7prueFmZn1p+vzSHulfXaeNbHsGvaj/c16rvTud38sG8/12h49BvU9rdLN0z/PBn0t/GrOby1l6H72W08bcGTwZMpMrnIwlOixRPo7ClR8gRU78kFz5DY/HK8P0oPw/YhKxGo94qCJ03xGFMgohy4os9ARH+BK1QdmJ3Ak9CTYyBtmaSoZMIUcFKVLLsCq5ctrJttxTW4MDdmqNTasH6OWVQCWyoRz5XAUiWy+srC5jigrIsxOiOnI5FQlTSZCpRFJar3LEzcq4pr+MLmPG6OlHPOWBVEIbAlEfHcCCxFIquu9IXNeYuy3k0fQGvldIAJWcBJVrIsC6xetrBzzzUnKWzO4+aMYucjWScqgS2ViOdKYKkSWX1lYXgBv63GaW+IyUpZyYQs4CQrWZYFVi0bCmMPVHGSUBhewOGZOB0u6OBlpWSiEnCqlCxXAvuv8vUX6rV2AXkIAAA=
-    http_version: 
-  recorded_at: Thu, 22 Jun 2017 18:57:53 GMT
-- request:
-    method: get
-    uri: https://capture.context.com/hawkular/metrics/status
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      User-Agent:
-      - rest-client/2.0.2 (linux-gnu x86_64) ruby/2.4.1p111
-      Hawkular-Tenant:
-      - default
-      Authorization:
-      - Bearer theToken
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Encoding:
-      - gzip
-      Connection:
-      - keep-alive
-      X-Powered-By:
-      - Undertow/1
-      Server:
-      - WildFly/10
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '154'
-      Date:
-      - Thu, 22 Jun 2017 18:57:54 GMT
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        H4sIAAAAAAAAAA3MvQ7CIBiF4XthFsJfK3Fr1aqDizTulH5NSIA2QF2M9y7TWZ7zftETSnI2a0gfZwGdkB6713i9oAN6hM1DgFhMcWvEb0i5biWU8JYwMrhofHX97nzBQ1oDvrmC9b1jFclmYlJZJSTMLadMLawRlgumjpOUojUgOVUL1MDZ5GzinEy97Rv6/QHfcrqrlgAAAA==
-    http_version: 
-  recorded_at: Thu, 22 Jun 2017 18:57:54 GMT
-- request:
-    method: get
-    uri: https://capture.context.com/hawkular/metrics/gauges/registry%2F476e8b1a-3a1a-11e7-a3eb-001a4a2314d7%2Fmemory%2Fusage/stats/?bucketDuration=20s&end=1498156842000&start=1498156542000
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      User-Agent:
-      - rest-client/2.0.2 (linux-gnu x86_64) ruby/2.4.1p111
-      Hawkular-Tenant:
-      - default
-      Authorization:
-      - Bearer theToken
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Encoding:
-      - gzip
-      Connection:
-      - keep-alive
-      X-Powered-By:
-      - Undertow/1
-      Server:
-      - WildFly/10
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '310'
-      Date:
-      - Thu, 22 Jun 2017 18:57:55 GMT
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        H4sIAAAAAAAAAM2SQWrDMBBF76K1KfLYGo297ylKF4K4JRCFEDulpeTuCdiWRvIUtKu3Dwu/N/y3XzVO7jqpvm47qg2aFrTWlRrOh8hwZv54Vj28EDSNBXq1lXJfnxnxw+Ho8s+8+87IePM5cf5yGsbnT58/95fpR/Uf7jQO92qjiIIiLWx+OV1v0kPaPkSdtFlTAzXA2hiJbRzObYwsbZwUt606iSIkip0mgx0yRUaiIoezIiOLIifliiAotgXnR2FaiPtqE6aFlCl2DRElioFwxQhXxUCCYiTlisKIrS44vxWmZfNp/W+bFaZl21SxBjDJQAJggpEtfgGsehGU2wn7tVhyeWFVlnaTJQyK9F7sSFgtQcHRSdgSbbaEmHetJAkLMJStJKYF8lfb+wMGsQ64bQcAAA==
-    http_version: 
-  recorded_at: Thu, 22 Jun 2017 18:57:55 GMT
-recorded_with: VCR 3.0.3
