+++ spec/vcr_cassettes/manageiq/providers/kubernetes/container_manager/metrics_capture/hawkular_capture_context_container_timespan.yml
@@ -26,8 +26,6 @@ http_interactions:
     headers:
       Content-Encoding:
       - gzip
-      Connection:
-      - keep-alive
       X-Powered-By:
       - Undertow/1
       Server:
@@ -37,16 +35,21 @@ http_interactions:
       Content-Length:
       - '154'
       Date:
-      - Thu, 22 Jun 2017 18:57:28 GMT
+      - Thu, 06 Jul 2017 10:58:17 GMT
+      Set-Cookie:
+      - a054b5d9e987bf679f10c9d29be39478=dca83a65476ba479e1cba709dfada9ed; path=/;
+        HttpOnly; Secure
+      Cache-Control:
+      - private
     body:
       encoding: ASCII-8BIT
       string: !binary |-
         H4sIAAAAAAAAAA3MvQ7CIBiF4XthFsJfK3Fr1aqDizTulH5NSIA2QF2M9y7TWZ7zftETSnI2a0gfZwGdkB6713i9oAN6hM1DgFhMcWvEb0i5biWU8JYwMrhofHX97nzBQ1oDvrmC9b1jFclmYlJZJSTMLadMLawRlgumjpOUojUgOVUL1MDZ5GzinEy97Rv6/QHfcrqrlgAAAA==
     http_version: 
-  recorded_at: Thu, 22 Jun 2017 18:57:28 GMT
+  recorded_at: Thu, 06 Jul 2017 10:58:17 GMT
 - request:
     method: get
-    uri: https://capture.context.com/hawkular/metrics/counters/registry%2F476e8b1a-3a1a-11e7-a3eb-001a4a2314d7%2Fcpu%2Fusage/stats/?bucketDuration=20s&start=1498156822000
+    uri: https://capture.context.com/hawkular/metrics/counters/registry%2F5bae6bfe-609c-11e7-9ac7-001a4a2314d5%2Fcpu%2Fusage/stats/?bucketDuration=20s&end=1499280042000&start=1499279722000
     body:
       encoding: US-ASCII
       string: ''
@@ -70,8 +73,6 @@ http_interactions:
     headers:
       Content-Encoding:
       - gzip
-      Connection:
-      - keep-alive
       X-Powered-By:
       - Undertow/1
       Server:
@@ -79,15 +80,20 @@ http_interactions:
       Content-Type:
       - application/json
       Content-Length:
-      - '1040'
+      - '426'
       Date:
-      - Thu, 22 Jun 2017 18:57:29 GMT
+      - Thu, 06 Jul 2017 10:58:17 GMT
+      Set-Cookie:
+      - a054b5d9e987bf679f10c9d29be39478=dca83a65476ba479e1cba709dfada9ed; path=/;
+        HttpOnly; Secure
+      Cache-Control:
+      - private
     body:
       encoding: ASCII-8BIT
       string: !binary |-
-        H4sIAAAAAAAAAI2ZXYtTMRRF/0ufB8nnOYnv/grxoeAoghVxRlHE/270Jmbv3gMefFu0w1rc3aTF1z8vT8/XL8+Xl7H0Fqu0lEIID5fHT283Kwe7ffg02IvcRaXFpq/6w+X67f2J3R7ffrieX3q7fj+xp6+3M7vePn98fBp8aNw+P/+4vHx3/fj0+OvhJFsMWSHZIll7ijmCLLEty/iQJTZlmfllxZBtkx3vfP7y1XpjO7+xB67sWUJOoWElMqgkPCuRrUpi7solRrKJZGuV2FVJltiWZXzIEpuyzPyyxth7cTySbgyv8/AkSOgqgsMjtisZH5XEZiUzf6UxvN5YtmoZ+9SEsshAlvCURbZkifllz2PXEP7/SPaL8I08PA05Bh2fXqgktisZH5XEZiUzb+U/MZLlU1Zzrqo9Z5RFBrKEpyyyJUvML3seuwbxPJLz8MZHlCvHJylGqRgJCBqRzkRAqxCRP9DYXOQDtpXSVLPg0yC2TRkfqsSmKzO3bDR2HpPjaURjc5E310OuKWXaHLFdyfioJDYrmfkrjc1FPmB7aTGmgVEWGcgSnrLIliwxv6yx8+i42feL4I2JhzfO+/E5EsVPCDGoJDwrka1KYu7KZAwv4QGbXoQ0gnII5Z/sHVuy9/iP7B37K3vP/LLG2JPjZt8vwjcKVzYdj672jpXIoJLwrES2Kon5K43hpUayMWrRlvCJINqqRA9TRFOUkN/T2Hn2XOrZ2FzmzUXJvRTpFQuRQSLh2YhsRRJzV2Zjc7mQbIolBFXFx0FsyzI+ZIlNWWZ+WWPn2XOpZ2NzmTeXak+9jCMcK5FBJeFZiWxVEvNXGsMrgWRz0FoCfk+/Y1uW8SFLbMoyc8sWY+zFc7MXY3iFh5drLBozPhFE0Ih0JgJahYj8gcbmCp+tY8t9fMOJKEpsmzI+VIlNV2Z+WWPnxXOpF2NzlTdXSpVSYs1YiQwqCc9KZKuSmLuyGpurfMDWUER7EDyziG1ZxocssSnLzC9r7Lx6LvVqDK/y8Gr+81NOpGElMqgkPCuRrUpi/kpjeJUPWAmx9TC+PoIssS3L+JAlNmWZ+WWNsYvnZhdjeMLDkzL+1ZTxZicGlYRnJbJVScxdKcbwhA9YURl/OJErIFBFOk0BLVFEfk9j5+K51MXYnPDmNEUd28TJIdqBRI9ARDOQkD/QmJvy2apjWuMXWRYURQamhKcqsuVKzC2rxsTVc5+rMTflubXcY2u14N6I7UrGRyWxWcnMX2lsTvlsbT230FLCr1jEQJbwlEW2ZIn5ZY2dq+dSV2N4jYfXUxkfo4q/ChHtRqJHIqJZSMgd2IzNNT5Wu9bcVDs+DWJgSniqIluuxPyyxs6b5z43/t9Hm3HOvfkNCPXxtN8aAAA=
+        H4sIAAAAAAAAAI2VQU7DMBBF75J1hGzHGc90zykQi0gNqFJTVU2KQIi7Y1Ri50+N8LJv4uj9+Nt9+mzmZbgszc56ERckOGeMaZvxtM/M39h0OEX2EKjvrLW9892jtW0zvL2W8DTuD0NxwTS8l/B8nYp4mM7HcY6jaDWdl49m9zIc5/GrvXP3BXfS7mRCsMF65Q4Y3HGS3AFnd8T17lRw5192W7lcrqWFfL+QjQpNlk3vmRyGRrwNrSZraMQptMLVoVdPcHfa3ceP6khEuQMGd5wkd8DZHXG9e+GgsK/YMC60lHVLiVm63vasQgOG0DhJoQHn0IjrQxdayqzcQ2/j64l6dEe8dVeT1R1xcle43r1wUMRUbJgUWiq6pUHitpI3VoUGDKFxkkIDzqERV4eWQktFX+exyiHeuUFd54i37mqyuiNO7grXuxcOilDNhhVaKrqlHKzzjli1FDGExkkKDTiHRlwf+q6lHH8r958HqaOgbgfEW3c1Wd0RJ3eFK92zJ7i7fzds89B2oW6pBGsoNkn9cSOG0DhJoQHn0Ij/Cv38DbMLHbEyCQAA
     http_version: 
-  recorded_at: Thu, 22 Jun 2017 18:57:29 GMT
+  recorded_at: Thu, 06 Jul 2017 10:58:17 GMT
 - request:
     method: get
     uri: https://capture.context.com/hawkular/metrics/status
@@ -114,8 +120,6 @@ http_interactions:
     headers:
       Content-Encoding:
       - gzip
-      Connection:
-      - keep-alive
       X-Powered-By:
       - Undertow/1
       Server:
@@ -125,16 +129,21 @@ http_interactions:
       Content-Length:
       - '154'
       Date:
-      - Thu, 22 Jun 2017 18:57:29 GMT
+      - Thu, 06 Jul 2017 10:58:17 GMT
+      Set-Cookie:
+      - a054b5d9e987bf679f10c9d29be39478=dca83a65476ba479e1cba709dfada9ed; path=/;
+        HttpOnly; Secure
+      Cache-Control:
+      - private
     body:
       encoding: ASCII-8BIT
       string: !binary |-
         H4sIAAAAAAAAAA3MvQ7CIBiF4XthFsJfK3Fr1aqDizTulH5NSIA2QF2M9y7TWZ7zftETSnI2a0gfZwGdkB6713i9oAN6hM1DgFhMcWvEb0i5biWU8JYwMrhofHX97nzBQ1oDvrmC9b1jFclmYlJZJSTMLadMLawRlgumjpOUojUgOVUL1MDZ5GzinEy97Rv6/QHfcrqrlgAAAA==
     http_version: 
-  recorded_at: Thu, 22 Jun 2017 18:57:30 GMT
+  recorded_at: Thu, 06 Jul 2017 10:58:17 GMT
 - request:
     method: get
-    uri: https://capture.context.com/hawkular/metrics/gauges/registry%2F476e8b1a-3a1a-11e7-a3eb-001a4a2314d7%2Fmemory%2Fusage/stats/?bucketDuration=20s&start=1498156842000
+    uri: https://capture.context.com/hawkular/metrics/gauges/registry%2F5bae6bfe-609c-11e7-9ac7-001a4a2314d5%2Fmemory%2Fusage/stats/?bucketDuration=20s&end=1499280042000&start=1499279742000
     body:
       encoding: US-ASCII
       string: ''
@@ -158,8 +167,6 @@ http_interactions:
     headers:
       Content-Encoding:
       - gzip
-      Connection:
-      - keep-alive
       X-Powered-By:
       - Undertow/1
       Server:
@@ -167,13 +174,18 @@ http_interactions:
       Content-Type:
       - application/json
       Content-Length:
-      - '748'
+      - '220'
       Date:
-      - Thu, 22 Jun 2017 18:57:30 GMT
+      - Thu, 06 Jul 2017 10:58:17 GMT
+      Set-Cookie:
+      - a054b5d9e987bf679f10c9d29be39478=dca83a65476ba479e1cba709dfada9ed; path=/;
+        HttpOnly; Secure
+      Cache-Control:
+      - private
     body:
       encoding: ASCII-8BIT
       string: !binary |-
-        H4sIAAAAAAAAAM2Zy2obQRBF/2XWwvSj+uV9viJkMWAlGDzGWHJICPn3KNF0PXoKUlppljpMo3ubQ9UIff41nc7z+3l69NCqT7lCcM4dpuPrE7F8Zcvz6/QYHpqPUKF+Kodp/v5tIMvx6XkeH1vmHwM5fSwjmZe3l+Pp8qWXL1/ezj+nx6/zy+n4+7CJmJWIdWXXk+f3D+1g3R5sblfdehwRMYwRawOQETsRERFixE4oIhJ7xKBEBMP1N0WtxtWKDyW14it1E6R3k/BvN0H+dZPE3k1Rq9V9RdxKXJz7//XTQ/xgkN2y86HyaghYM2JrMQS9FwFrLUwi0sFu0m3VLS5bLn0rVHF1rNWgBNmrE1EMITbrhKohsXdThPJORKwRckk8IiMUkcNrREbWiJyYI3pFXR8M1+8VqzwM3ZLjE18Q3o1g74YEuxGxd1PU8nlfERWJvWHh0kPsYBjVSsC3mSCiG/BtJgh1g1u3GcYREcMYsbkh4QpEwM4w3wooXgf2dIq/wbBr6SF+cLQqRx+z6IWEFyPYmyHBakTs3RSrghyNNZeaxMhnhEVkcI1IpEdkxB5R8Tdadm1UrIqDVblebkBoRYR3I9i7IcFuRMzdoqJWhH1FVCSOlq0bFbXiqNaduylqgdtVRFAkBsvWBUUt2KiVxTsPAdEsizceAtQr3/q+A4pVMIzGy++w6sVoJMLzEewBkWBCIvaIir9gWbigWJVGqxo0L68eieiGELt1Qt2QmLslxao0jsbWnJNjH4mIiBAjdkIRkdgjKv4my9ZNilppo9Z9uylqpXE03jmiInG2bN2sqJUHtYqLXr7MEWHdGFy7EendGDF3y4paGcaIcjL2zyKenIv9M0W7dSpmRd1s2bVZESoPQhXvQxPjlAhvRbD3QoLNiNi7KUIVt6uIRVG3WHZtUYQqo1B37qaoVfK+IioSF8vCLYpadV9qVUWtOk7FO0dUJK6Whav8W1GqMrW+/AFCRjlq6BgAAA==
+        H4sIAAAAAAAAAM2SQQrCMBBF75J1kZiGZqZ7TyEuAo1SaEppoiji3S1UaZsO2J2zzGc+8zK841OEaPsoyr1GVAaNVlLKTLi2mrJizHzdijLf5bmBXOqDyYS9XZLEu6q26Zi39yQJV58m1neNC8PSYbnv4kOUZ9sE98pWiAWBCJ9sbMb+ShVhXQTJ6m9fnAWi4oWoCES94fxAqAW81AJCLQBeiITEKDecHwm1kJdaSKiFmhciITEWW85PqIW81MKVWjC8GSFOOAtE9fP8s6F58a9qnd5t1VyWfAcAAA==
     http_version: 
-  recorded_at: Thu, 22 Jun 2017 18:57:30 GMT
+  recorded_at: Thu, 06 Jul 2017 10:58:17 GMT
 recorded_with: VCR 3.0.3
